rename user.ts to user.model.ts
